Add yml templates that replace the md ones and config.yml
This adds forms to our bug and feature reports. As well it specifies a config.yml that is shown when users click to create
a feature or bug report. Also adds a SUPPORT.md as specified under the community health files:
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/creating-a-default-community-health-file
This lets Users know where to get support. I'd recommend in the future if we can set it up, adding a FUNDING.yml
eventually.