Create guides/release-checklist.md
This guide is an inventory and step-by-step checklist of things
that need doing when making a Regular release of Pulsar over at the
Pulsar core repo, as well as some additional how-to style info for
one or two of the steps.

This may eventually need to be updated, but having this linkable
somewhere prominent will be a good thing, so posting it as it is now.

This is imported from the gist this was in before, and then lightly
edited/improved before being committed here.