Create rules-info-channel.md
This is a update to the rules(and potentionally a renaming to rules-and-info). This better organizes that channel and gets rid of unnecessary messages that were pings to make sure people saw things in there and remembered to join roles. As well as adding one and changing the description of the other.

I'd also like to propose this as the starting point to also discuss the colors of those roles for accessibility purposes